pyproject.toml
[tool.poetry]
name = "cloud_computing"
version = "0.1.0"
description = ""
authors = ["Mitar Perovic <perovicmitar@uns.ac.rs>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "=3.2.*"
djangorestframework = "=3.12.*"
serializers = "^0.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
